Prevent memory-cleanup on exit.  This leads in some case to an stall. So disable it for now. Speed-up thread routines for getting self memory-address and cleanup some code.
git-svn-id: https://mingw-w64.svn.sourceforge.net/svnroot/mingw-w64/experimental/winpthreads@4395 4407c894-4637-0410-b4f5-ada5f102cad1

+++ src/sem.h
@@ -34,7 +34,7 @@ struct _sem_t
 {
     unsigned int valid;
     HANDLE s;
-    volatile int value;
+    volatile long value;
     pthread_mutex_t vlock;
 };
